+++ src/app/(route)/(withLayout)/votes/[voteId]/_component/VoteInfo.tsx
@@ -1,120 +1,60 @@
 'use client'
 
-import { useCallback, useState } from 'react'
-import { useParticipationVote } from '@/app/_hook/api/votes/useParticipationVote'
-import { VoteDetailType } from '@/app/_types/detailVote.type'
-import { useReParticipation } from '@/app/_hook/api/votes/useReParticipation'
+import { categoryFormatter } from '@/app/_utils/categoryFormatter'
+import { dateFormatter } from '@/app/_utils/dateFormatter'
 import { cn } from '@/app/_utils/twMerge'
-import ProgressBar from './ProgressBar'
-import VoteItem from './FavoritesVoteItem'
+import { VoteDetailType } from '@/app/_types/detailVote.type'
 import VoteProgressTracker from './VoteProgressTracker'
-import { validateChoiceItem } from '../_utils/validation'
+import ManagementButton from './ManagementButton'
 
 export default function VoteInfo({ voteData }: { voteData: VoteDetailType }) {
-  const { item1Info, item2Info, voteInfo, selectedItemId } = voteData
+  const { isOwner, voteInfo } = voteData
   const {
-    id: voteId,
-    item1Votes,
-    item2Votes,
-    isVoting,
-    participants,
+    content,
+    startTime,
     endTime,
+    maxParticipants,
+    participants,
+    id,
+    hobby,
   } = voteInfo
-
-  const [itemId, setItemId] = useState<number | null>(null)
-
-  const { mutateAsync: voteItem } = useParticipationVote()
-  const { mutateAsync: reVote } = useReParticipation()
-
-  const handleSelectItem = useCallback(
-    (selectItemId: number) => {
-      if (itemId === selectItemId) {
-        setItemId(null)
-      } else {
-        setItemId(selectItemId)
-      }
-    },
-    [itemId, setItemId, selectedItemId],
-  )
-
-  /** 투표 참여 */
-  const handleVote = async () => {
-    if (!validateChoiceItem(itemId)) {
-      return
-    }
-
-    const status = await voteItem({ itemId, voteId })
-
-    if (status === 200) {
-      setItemId(null)
-    }
-  }
-
-  /** 재투표(투표 취소) */
-  const handleReVote = async () => {
-    await reVote(voteId)
-  }
 
   return (
     <article
       className={cn(
-        'mt-[12px] flex h-[567px] items-center rounded-[8px] border border-[#E6E6E6]  px-[106px]',
-        'mo:mt-0 mo:h-[520px] mo:w-full mo:items-start mo:border-0 mo:px-[16px]',
+        'mt-[21px] rounded-[8px] border border-[#E6E6E6] px-[40px] py-[30px]',
+        ' mo:w-full mo:border-0 mo:px-[16px]',
       )}
     >
-      <div className={cn('h-[464px] w-[514px]', 'mo:w-full')}>
-        <div
-          className={cn(
-            'flex h-[338px] justify-between',
-            'mo:justify-center mo:gap-[8px]',
-          )}
-        >
-          {/** item1 */}
-          <VoteItem
-            voteItemInfo={item1Info}
-            itemId={itemId}
-            onSelectItem={handleSelectItem}
-          />
+      <div className="flex justify-between text-center text-[12px] font-[500]">
+        <div className="flex gap-[6px]">
           <div
             className={cn(
-              'mt-[16px] h-[262px] w-[0.77px] bg-[#DEDEDE]',
-              'mo:hidden',
+              'h-[24px] rounded-[8px] bg-[#F2F2F2] px-[6px] py-[4px]',
+              'mo:bg-black mo:text-white',
             )}
-          />
-          {/** item2 */}
-          <VoteItem
-            voteItemInfo={item2Info}
-            itemId={itemId}
-            onSelectItem={handleSelectItem}
-          />
+          >
+            {categoryFormatter(hobby)}
+          </div>
+          <div
+            className={cn(
+              'h-[24px] rounded-[8px] bg-[#F2F2F2] px-[6px] py-[4px]',
+              'mo:bg-black mo:text-white',
+            )}
+          >
+            {hobby}
+          </div>
         </div>
-        <div className={cn('hidden', 'mo:block mo:pt-[16px]')}>
-          <VoteProgressTracker endTime={endTime} participants={participants} />
-        </div>
-        {/** 투표 게이지 */}
-        {(selectedItemId || !isVoting) && (
-          <ProgressBar item1Votes={item1Votes} item2Votes={item2Votes} />
-        )}
-        <div className={cn('mt-[36px] flex justify-center', '')}>
-          {isVoting && (
-            <button
-              type="button"
-              className={cn(
-                'h-[48px] w-[136px] rounded-[100px] font-[600] text-[#fff]',
-                'mo:w-full mo:rounded-[4px]',
-                {
-                  'bg-[#000]': itemId,
-                  'bg-[#757575]': !itemId,
-                  'px-[40px]': !selectedItemId,
-                  'px-[22px]': selectedItemId,
-                },
-              )}
-              onClick={selectedItemId ? handleReVote : handleVote}
-            >
-              {selectedItemId ? '투표 취소하기' : '투표하기'}
-            </button>
-          )}
-        </div>
+        {isOwner && <ManagementButton voteId={id} />}
+      </div>
+      <div className={cn('mt-[18px] flex', 'mo:mt-[8px]')}>
+        <span className="flex h-[17px] gap-[4px] text-[12px] font-[500] text-[#747474]">
+          {dateFormatter(startTime)} · 투표인원 {maxParticipants}명
+        </span>
+      </div>
+      <p className="mt-[26px] text-[14px] font-[500]">{content}</p>
+      <div className={cn('block', 'mo:hidden')}>
+        <VoteProgressTracker endTime={endTime} participants={participants} />
       </div>
     </article>
   )
